refactor the generation into engine
