+++ README.md
@@ -1,2 +1,7 @@
 # jaxml
-Some testing ground of jax
+
+A relatively carefully written library to support my own research (and hopefully help others).
+
+# Llama model inference
+
+The first
