+++ src/jaxml/models/llama.py
@@ -18,7 +18,6 @@ from typing import Any, Optional
 import flax
 import jax
 import jax.numpy as jnp
-from jax.experimental import mesh_utils
 from jax.sharding import Mesh, NamedSharding, PartitionSpec
 from flax import linen as nn
 from flax.linen.partitioning import with_sharding_constraint
@@ -205,135 +204,10 @@ class LlamaModel(Block):
         )        
         
         
-class LlamaForCausalLM(Block):
+class LlamaModelWithHead(Block):
     
     lm_head_init: Any = nn.initializers.xavier_uniform
     lm_head_init_args: tuple = ()
-
-    def init_cache(self, max_seq_len) -> list[KVCache]:
-        num_layers = self.config.num_layers
-        return [KVCache.init(max_seq_len, None, None, dtype=self.dtype) for _ in range(num_layers)]
-
-    @staticmethod
-    def mesh_sharding(pspec: Optional[PartitionSpec], mesh: Optional[Mesh]) -> NamedSharding:
-        if mesh is None:
-            mesh = Mesh(jax.devices(), (None,))
-        return NamedSharding(mesh, pspec)
-
-    def _shard_params(self, x, y: PartitionSpec):
-        if x.ndim != len(y.spec):
-            assert (
-                x.ndim == 2 and len(y.spec) == 3
-            ), f"The shape of x ({x.shape}) and the sharding spec ({y.spec}) does not match"
-            warnings.warn(
-                f"The parameter has 2 axis ({x.shape}) while the sharding spec ({y.spec}) has 3 axis. "
-                "Attempting to reshape into [:, :, head_dim], but please confirm that this is the intended behavior."
-            )
-            return jax.device_put(
-                x.reshape(
-                    (
-                        x.shape[0],
-                        -1,
-                        self.head_dim,
-                    )
-                ),
-                y,
-            )
-        return jax.device_put(x, y)
-
-
-    def get_params(self, tp_size: int = 4, weights: Any = None, sharded: bool = True):
-        """
-        Get the properly sharded parameters.
-        Args:
-            tp_size: the tensor-parallel size.
-            weights: whether a tree of weights are already given (but may not be sharded).
-        Returns:
-            a tree of properly sharded parameters
-        """
-        if sharded and not (len(jax.devices()) > 1 and len(jax.devices()) % tp_size == 0):
-            sharded = False
-            warning.warn(f"Cannot shard across devices: {jax.devices()}.")
-
-        key = jax.random.PRNGKey(0)
-
-        # (dp, tp)
-        mesh_layout = (len(jax.devices()) // tp_size, tp_size)
-
-        dummy_input = jnp.array(
-            [[1 for _ in range(mesh_layout[1])] for _ in range(mesh_layout[0])]
-        )
-        abstract_variables = jax.eval_shape(self.init, key, dummy_input)
-        rules = (
-            ("batch", "data"),
-            ("heads", "model"),
-            ("kv_length", None),
-            ("length", None),
-            ("intermediate", "model"),
-            ("heads_merged", "model"),
-            ("head_states", None),
-        )
-
-        if sharded:
-            mesh = Mesh(
-                devices=mesh_utils.create_device_mesh(mesh_layout),
-                axis_names=("data", "model"),
-            )
-
-            logical_state_spec = nn.get_partition_spec(abstract_variables)
-            logical_state_sharding = nn.logical_to_mesh_sharding(
-                logical_state_spec, mesh, rules
-            )
-
-            input_sharding = self.mesh_sharding(
-                PartitionSpec("data", None), mesh
-            )  # dimensions: (batch, length)
-
-            if weights is not None:
-                assert isinstance(
-                    weights, dict
-                ), f"weights must be a dict, got {type(weights)}"
-                assert (
-                    "params" in weights
-                ), f"The key params not found in 'weights'. Got {weights.keys()}"
-
-                if sharded:
-                    params = {
-                        "params": jax.tree_util.tree_map(
-                            self._shard_params,
-                            weights["params"],
-                            logical_state_sharding["params"],
-                        )
-                    }
-                else:
-                    params = weights
-            else:
-                params = jax.jit(
-                    self.init,
-                    in_shardings=(
-                        self.mesh_sharding(None, mesh),
-                        input_sharding,
-                    ),  # PRNG key and x
-                    out_shardings=logical_state_sharding,
-                )(key, dummy_input)
-        else:
-            params = self.init(key, dummy_input)
-
-        return params
-
-    def prepare_input(self, inputs, dp_size: int = 1, dtype: Any = None):
-        mesh = Mesh(
-            devices=mesh_utils.create_device_mesh(
-                (dp_size, jax.device_count() // dp_size)
-            ),
-            axis_names=("data", "model"),
-        )
-        inputs = jax.device_put(
-            inputs, self.mesh_sharding(PartitionSpec("data", None), mesh)
-        )
-        if dtype is not None:
-            inputs = jax.tree_util.tree_map(lambda x: x.astype(dtype), inputs)
-        return inputs
 
     def setup(self):
         self.model = LlamaModel(
@@ -375,65 +249,3 @@ class LlamaForCausalLM(Block):
             hidden_states=outputs.hidden_states,
             attention_weights=outputs.attention_weights,
         )
-
-    def wrapped_apply_fn(
-        self,
-        params,
-        input_token,
-        attention_mask=None,
-        kv_caches=None,
-        use_cache=True,
-    ) -> tuple[jnp.ndarray, list[KVCache]]:
-
-        out, _ = self.apply(
-            params,
-            input_token,
-            position_ids=None,
-            attention_mask=attention_mask,
-            mutable=("cache",),
-            #output_hidden_states=False, # maybe allow for toggling of hidden states in the future
-            #output_attentions=False, # maybe allow for toggling of attn wts in the future
-            kv_caches=kv_caches,
-            use_cache=use_cache,
-        )  # return a tuple (CausalLMOutputWithCache, dict) where dict is the mutable cache
-
-        return out.logits, out.kv_caches
-
-    def generate(
-        self,
-        params,
-        prompt_tokens: jnp.ndarray,
-        attention_mask: Optional[jnp.ndarray] = None,
-        do_sample: bool = True,
-        seed: int = 0,
-        max_new_tokens: int = 10,
-        top_k: int = 0,
-        top_p: float = 0.0,
-        temperature: float = 1.0,
-        no_jit: bool = False,
-        show_progress: bool = False,
-    ):
-        if no_jit:
-            apply = self.wrapped_apply_fn
-        else:
-            apply = jax.jit(self.wrapped_apply_fn, static_argnames=("use_cache",))
-
-        kv_caches = self.init_cache(max_seq_len=prompt_tokens.shape[1] + max_new_tokens)
-        
-        from .._generate import generate
-
-        return generate(
-            params,
-            apply,
-            prompt_tokens,
-            attention_mask,
-            kv_caches,
-            do_sample=do_sample,
-            seed=seed,
-            max_new_tokens=max_new_tokens,
-            top_k=top_k,
-            top_p=top_p,
-            temperature=temperature,
-            show_progress=show_progress,
-        )
-
