+++ tests/test_llama.py
@@ -19,6 +19,7 @@ import torch
 
 from jaxml.utils import torch_to_jax_states
 from jaxml.cache import KVCache
+from jaxml.inference_engine.engine import Engine, InferenceConfig
 
 
 def test_llama_mlp(llama_mlp, hf_llama_mlp):
@@ -77,24 +78,28 @@ def test_llama_model(llama_model, hf_llama_model):
             )
         
         assert np.allclose(y.last_hidden_state, y2.last_hidden_state.numpy(), atol=1e-5)
-        #print(len(y.attention_weights), [type(t) for t in y2[1]])
         assert all(np.allclose(a, b.numpy(), atol=1e-5) for a, b in zip(y.attention_weights, y2.attentions))
         assert all(np.allclose(a, b.numpy(), atol=1e-5) for a, b in zip(y.hidden_states, y2.hidden_states))
 
-def test_llama_causal_model(llama_causal_model, hf_llama_causal_model):
+def test_llama_completion(llama_model_with_head, hf_llama_causal_model):
     bs, seq_len = 4, 10
     with jax.default_device(jax.devices("cpu")[0]):
-        model, init_param = llama_causal_model
+        model, init_param = llama_model_with_head
         key = jax.random.PRNGKey(0)
         x = jax.random.randint(key, (bs, seq_len), 0, model.config.vocab_size - 1, dtype=jnp.int32)
         params = torch_to_jax_states(hf_llama_causal_model, head_dim=model.head_dim, dtype=torch.float32)
-        y = model.generate({**init_param, "params": params["params"]}, x, max_new_tokens=10, temperature=0.0)
+        
+        config = InferenceConfig()
+        engine = Engine(model, config, {**init_param, "params": params["params"]})
+        engine.init_params(use_tpu=False)
+        y = engine.generate(x, max_new_tokens=10, temperature=0.0)
         with torch.no_grad():
             y2 = hf_llama_causal_model.generate(
                 input_ids=torch.tensor(np.array(x)),
                 max_new_tokens=10,
                 do_sample=False,
             )
+        print(y, y2)
         assert np.all(y == y2.numpy())
 
 def test_kv_cache(llama_model):
